chore: refactor more of the UI into its own file and adjust the chart styling
