+++ web.go
@@ -28,18 +28,11 @@ type chartProps struct {
 }
 
 var charts = []chartProps{
-	{"TPS", "Throotle", nil},
-	{"RPM", "Revvie wevvy", nil},
+	{"TPS", "Throttle", nil},
+	{"RPM", "Revolutions Per Minute", nil},
 }
 
-func buildUpdateChartScript(name string, x, y int) string {
-	return fmt.Sprintf(`pushData("%sBuffer", %d, %d);`,
-		strings.ToLower(name),
-		x, // should already be ms since epoch
-		y,
-	)
-}
-
+// IndexHandler is the main entrypoint for the UI
 func IndexHandler(w http.ResponseWriter, _ *http.Request) {
 	err := Templates.ExecuteTemplate(w, "index", map[string]interface{}{
 		"cards": cards,
@@ -55,10 +48,13 @@ func IndexHandler(w http.ResponseWriter, _ *http.Request) {
 		},
 	})
 	if err != nil {
+		fmt.Println(err)
 		w.WriteHeader(http.StatusInternalServerError)
 	}
 }
 
+// EventsHandler is called on page load and pushes page changes to the client via SSE,
+// based on events generated by the Huskki input source (live / replay)
 func EventsHandler(w http.ResponseWriter, r *http.Request) {
 	sse := ds.NewSSE(w, r)
 
@@ -78,5 +74,71 @@ func EventsHandler(w http.ResponseWriter, r *http.Request) {
 				return
 			}
 		}
+	}
+}
+
+func buildUpdateChartScript(name string, x, y int) string {
+	return fmt.Sprintf(`pushData("%s", %d, %d);`, strings.ToLower(name), x, y)
+}
+
+// generatePatch takes an event received from the event queue, iterates the cards that are displayed on the UI,
+// and returns a closure that can be used to patch the client.
+func generatePatch(event map[string]any) func(*ds.ServerSentEventGenerator) error {
+
+	var writer = strings.Builder{}
+	var funcs []func(generator *ds.ServerSentEventGenerator) error
+
+	// For each card, see if we have an update and template a response
+	for _, card := range cards {
+		if value, ok := event[strings.ToLower(card.Name)]; ok {
+			Templates.ExecuteTemplate(&writer, "card.value", cardProps{Name: card.Name, Value: fmt.Sprintf("%v", value)})
+		}
+	}
+
+	// For each chart see if we have an update and form an SSE update function
+	for _, chart := range charts {
+		value, ok := event[strings.ToLower(chart.Name)]
+		if !ok {
+			continue
+		}
+		timestamp, ok := event["timestamp"]
+		if !ok {
+			continue
+		}
+
+		v, ok := value.(int)
+		if !ok {
+			continue
+		}
+		ts, ok := timestamp.(int)
+		if !ok {
+			continue
+		}
+
+		funcs = append(funcs, func(sse *ds.ServerSentEventGenerator) error {
+			err := sse.ExecuteScript(buildUpdateChartScript(chart.Name, ts, v))
+			return err
+		})
+	}
+
+	// Main closure
+	return func(sse *ds.ServerSentEventGenerator) error {
+		// Patch UI elements
+		if writer.String() != "" {
+			err := sse.PatchElements(writer.String())
+			if err != nil {
+				return err
+			}
+		}
+
+		// Exec client-side javascript
+		for _, f := range funcs {
+			err := f(sse)
+			if err != nil {
+				return err
+			}
+		}
+
+		return nil
 	}
 }
